moved todo list component back to App.js for cohesive local storage update.
